// src/lib/api/models/InvoiceDeliveryState.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Enum for filtering invoices by delivery date state.
 */
export enum InvoiceDeliveryState {
    ALL = 'all',
    SCHEDULED = 'scheduled',
    UNSCHEDULED = 'unscheduled',
}