// src/lib/api/models/OrderBillingStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Enum for filtering orders by billing status.
 */
export enum OrderBillingStatus {
    ALL = 'all',
    FULLY_BILLED = 'fully_billed',
    NOT_FULLY_BILLED = 'not_fully_billed',
}