// src/lib/api/models/CommissionPeriodResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CommissionPeriodResponse = {
    id: number;
    year: number;
    month: number;
    day: number;
    total_commission: number;
    created: number;
    updated: (number | null);
};